+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/internal/audio/AudioClientFactory.kt
@@ -24,6 +24,7 @@ class AudioClientFactory private constructor(
             audioClientObserver,
             audioClientObserver,
             audioClientObserver,
+            audioClientObserver,
             0
         )
     }